criado  controller de upload de imagens no sistema

+++ ProEventos.Api/Controllers/EventoController.cs
@@ -1,6 +1,8 @@
 ﻿using System;
-
+using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
@@ -14,13 +16,12 @@ namespace ProEventos.Api.Controllers
     public class EventoController : ControllerBase
     {
         private readonly IEventoService _eventoService;
+        private readonly IWebHostEnvironment _hostEnvironment;
 
-
-
-
-        public EventoController(IEventoService eventoService)
+        public EventoController(IEventoService eventoService,IWebHostEnvironment hostEnvironment)
         {
             _eventoService = eventoService;
+            _hostEnvironment = hostEnvironment;
         }
 
         [HttpGet]
@@ -32,12 +33,14 @@ namespace ProEventos.Api.Controllers
                 if (eventos == null) return NoContent();
 
                 return Ok(eventos);
+
             }
             catch (Exception e)
             {
                 return this.StatusCode(StatusCodes.Status500InternalServerError,
                     $"Erro ao tentar recuperar eventos. Erro{e.Message}");
             }
+            
         }
 
         [HttpGet("{id}")]
@@ -97,6 +100,31 @@ namespace ProEventos.Api.Controllers
             }
             
         }
+        [HttpPost("image/{eventoId}")]
+         public async Task<IActionResult> UploadImage(int eventoId)
+        {
+            try
+            {
+                var evento = await _eventoService.GetEventosByIdAsync(eventoId);
+                if (evento == null)
+                    return NoContent();
+
+                var file  = Request.Form.Files[0];
+                if(file.Length > 0){
+                    DeleteImage(evento.ImageUrl);
+                    evento.ImageUrl = await SaveImage(file);
+                }
+                var EventoRetorn  = await _eventoService.UpdateEvento(eventoId,evento);
+
+                return Ok(evento);
+            }
+            catch (Exception e)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError,
+                    $"Erro ao tentar adicionar imagem. Erro{e.Message}");
+            }
+            
+        }
         
         [HttpPut("{id}")]
         public async Task<IActionResult> Put(int id, EventoDto model)
@@ -121,9 +149,12 @@ namespace ProEventos.Api.Controllers
         {
             try
             {
+                EventoDto evento = await  _eventoService.GetEventosByIdAsync(id,true);
+                if(evento == null) return NoContent();
 
                 if (await _eventoService.DeleteEvento(id))
                 {
+                    DeleteImage(evento.ImageUrl);
                     return Ok("Deletado");
                 }else
                     return BadRequest("evento nao deletado");
@@ -136,5 +167,31 @@ namespace ProEventos.Api.Controllers
             }
             
         }
+
+        [NonAction]
+        public void DeleteImage(string imageName){
+            var imagePath  = Path.Combine(_hostEnvironment.ContentRootPath, @"Resources/images", imageName);
+            if(System.IO.File.Exists(imagePath)){
+                System.IO.File.Delete(imagePath);
+            }
+
+        }
+
+         [NonAction]
+        public async Task<string> SaveImage(IFormFile imageFile){
+           var imageName = new string(Path.GetFileNameWithoutExtension(imageFile.FileName)
+                                            .Take(15).ToArray()
+                                            ).Replace(' ', '-');
+
+           imageName = $"{imageName}{DateTime.UtcNow.ToString("yymmssfff")}{Path.GetExtension(imageFile.FileName)}";
+
+           var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"Resources/images", imageName);
+           using(var fileStream = new FileStream(imagePath,FileMode.Create)){
+                await imageFile.CopyToAsync(fileStream);
+           }
+            return imageName;
+
+        }
+
     }
 }
